Extract instruction cache to cache.go
